@@ -2,7 +2,8 @@
   "name": "lorecharacters",
   "private": true,
   "dependencies": {
-    "@rails/webpacker": "^3.5.5",
+    "@babel/runtime": "^7.7.2",
+    "@rails/webpacker": "^4.0.7",
     "axios": "^0.18.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-2": "^6.24.1",
@@ -11,29 +12,40 @@
     "coffeescript": "1.12.7",
     "jest": "^23.6.0",
     "jquery": "^3.3.1",
+    "mini-css-extract-plugin": "^0.8.0",
     "moment": "^2.22.2",
     "moxios": "^0.4.0",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
     "react-redux": "^5.1.1",
+    "react-scripts": "3.2.0",
     "react-test-renderer": "^16.6.3",
     "react_ujs": "^2.4.4",
     "reactstrap": "^6.5.0",
     "redux": "^4.0.1",
     "redux-mock-store": "^1.5.3",
     "redux-thunk": "^2.3.0",
-    "webpack-cli": "^3.1.2",
+    "webpack-cli": "^3.3.9",
     "webpacker-react": "^0.3.2"
   },
   "devDependencies": {
-    "webpack-dev-server": "^2.11.3"
+    "@babel/core": "^7.7.2",
+    "@babel/helper-plugin-utils": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.7.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.1.6",
+    "@babel/preset-react": "^7.7.0",
+    "babel-jest": "^24.9.0",
+    "regenerator-runtime": "^0.13.3",
+    "webpack-dev-server": "^3.9.0"
   },
   "jest": {
     "testMatch": [
       "**/__tests__/**/*.js?(x)"
     ],
     "transform": {
-      "^.+\\.(js|jsx)$": "<rootDir>/node_modules/babel-jest"
+      "^.+\\.(js|jsx)$": "babel-jest"
     },
     "transformIgnorePatterns": [
       "node_modules/(?!react-native|react-navigation)"
