@@ -4,45 +4,11 @@ import {handleFormChange, signIn, switchForms, submitRegistration} from './actio
 import {Container} from 'reactstrap';
 import AccountForms from './components/accountForms'
 
-const mapStateToProps = (
-    state
-) => {
-  return {
-    ...state
-  }
-};
-
-const mapDispatchToProps = (
-    dispatch
-) => {
-  return {
-    handleFormChange: (fieldName, fieldValue) => {
-      dispatch(handleFormChange(fieldName, fieldValue))
-    },
-    signIn: () => {
-      dispatch(signIn())
-    },
-    submitRegistration: () => {
-      dispatch(submitRegistration())
-    },
-    switchForms: (formId) => {
-      dispatch(switchForms(formId))
-    }
-  }
-};
-
-const AccountFormsComponent = connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(AccountForms);
-
-
-
 export default class SignInContainer extends React.Component {
   render () {
     return(
         <Container>
-          <AccountFormsComponent/>
+          <AccountForms/>
         </Container>
     )
   }
